requirements.txt
absl-py==0.13.0
astor==0.8.1
biopython==1.79
cached-property==1.5.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.39.0
h5py==2.10.0
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==2.6.8
networkx==2.6.2
numpy==1.16.5
opt-einsum==3.3.0
pandas==1.2.0
protobuf==4.0.0rc2
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
tensorboard==1.15.0
tensorflow==1.15.5
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.2.0
Werkzeug==0.11.15
wrapt==1.13.0rc2
